move check_duplicates to the access_helpers
